Fixing type mismatch in Listing 3-18

+++ chapter_3/listing_3_18.cpp
@@ -25,7 +25,7 @@ private:
 };
 
 struct Avout {
-  Avout(const char* name, long year_of_apert)
+  Avout(const char* name, int year_of_apert)
     : name{ name }, apert{ year_of_apert } {}
   void announce() const {
       printf("My name is %s and my next apert is %d.\n", name, apert.get_year());
